+++ field/eisenstein/eisenstein.go
@@ -2,7 +2,7 @@
 // algebraic number field Q(ω) – the third cyclotomic field.  These are of the
 // form z = a + bω, where a and b are integers and ω is a primitive third root
 // of unity i.e. ω²+ω+1 = 0.
-package ecc
+package eisenstein
 
 import (
 	"math/big"
