+++ field/eisenstein/eisenstein_test.go
@@ -1,4 +1,4 @@
-package ecc
+package eisenstein
 
 import (
 	"crypto/rand"
